+++ src/main.rs
@@ -150,6 +150,11 @@ Note: If qps is specified, burst will be ignored",
         long = "unix-socket"
     )]
     unix_socket: Option<std::path::PathBuf>,
+    #[clap(
+        help = "Include a response status code successful or not successful breakdown for the time histogram and distribution statistics",
+        long = "stats-success-breakdown"
+    )]
+    stats_success_breakdown: bool,
 }
 
 /// An entry specified by `connect-to` to override DNS resolution and default
@@ -367,7 +372,7 @@ async fn main() -> anyhow::Result<()> {
                         }
                         _ = ctrl_c_rx.recv_async() => {
                             // User pressed ctrl-c.
-                            let _ = printer::print_result(&mut std::io::stdout(),print_mode,start, &all, start.elapsed(), opts.disable_color);
+                            let _ = printer::print_result(&mut std::io::stdout(),print_mode,start, &all, start.elapsed(), opts.disable_color, opts.stats_success_breakdown);
                             std::process::exit(libc::EXIT_SUCCESS);
                         }
                     }
@@ -390,6 +395,7 @@ async fn main() -> anyhow::Result<()> {
                 start,
                 fps: opts.fps,
                 disable_color: opts.disable_color,
+                stats_success_breakdown: opts.stats_success_breakdown,
             }
             .monitor(),
         )
@@ -565,6 +571,7 @@ async fn main() -> anyhow::Result<()> {
         &res,
         duration,
         opts.disable_color,
+        opts.stats_success_breakdown,
     )?;
 
     Ok(())