+++ src/monitor.rs
@@ -55,6 +55,7 @@ pub struct Monitor {
     // Frame per scond of TUI
     pub fps: usize,
     pub disable_color: bool,
+    pub stats_success_breakdown: bool,
 }
 
 impl Monitor {
@@ -452,6 +453,7 @@ impl Monitor {
                             &all,
                             now - self.start,
                             self.disable_color,
+                            self.stats_success_breakdown,
                         );
                         std::process::exit(libc::EXIT_SUCCESS);
                     }